Fix half precision

+++ benchmarks/product_matcsrmat-bench.cpp
@@ -94,7 +94,7 @@ void run(PreferredBackend backend)
   }
 
   MemoryPool::synchronize();
-  std::cout<<"control norm: "<<r.norm_frobenius()<<"\n";
+  std::cout<<"control norm: "<<double(r.norm_frobenius())<<"\n";
 }
 
 int main(int argc, char ** argv)
